AI

Description

This module defines how the AI plays. applyDifficulty, boardScore, and playHand are exported for tests

Documentation

applyDifficulty :: forall a. Ord a => Difficulty -> StdGen -> [a] -> [[a]]

Given the hand, the permutations to consider for playing this hand

play

Arguments

:: Difficulty

-> SharedModel

-> Board Core

-> Player
The playing player

-> [Place]
Events generated for player pSpot

Smart play events

boardPlayerScore :: Board Core -> Player -> Nat

The score of given player's in-place cards. Smaller is the best. Both negative and positive values are returned.

data Difficulty

The AI's level

Constructors

Easy

Medium

Hard

placeCards

Arguments

:: Difficulty

-> SharedModel

-> Board Core

-> Player
The player whose cards must be played

-> [Place]

Events that place creatures on the board. This function guarantees that the returned events are solely placements (no neutral cards), so that playing them with playAll returns an empty list of Event FIXME @smelc remove moe, use play directly

playHand

Arguments

:: SharedModel

-> Board Core

-> Player
The playing player

-> [Place]

Events for playing all cards of the hand, in order. Each card is played optimally.