Art

Description

Module defining a function Board -> ASCII. Not in Board itself, because it depends on Total and Total needs to depend on Board. |

Documentation

toASCII :: Board Core -> String

The spot of a card, as visible from the top of the screen. For the