Board

Documentation

appliesTo :: ID -> T Core -> Player -> Card -> Bool

data Teams a

Constructors

Teams

Fields

topTeam :: a

botTeam :: a

data InPlaceEffect

Initially this type was for displaying animations only. However Game.hs also uses for Core stuff internally (see applyInPlaceEffectOnBoard). Unfortunate :-( So be careful when changing related code.

Constructors

InPlaceEffect

Fields

attackChange :: Int

Attack value changed

death :: DeathCause

Did creature die? If yes, for what reason

attackBump :: Bool

Creature attacked (value used solely for animations)

hitPointsChange :: Int

Hits points changed

extra :: Nat

Hit points that were not dealt because the defender died. Contributes to the score with Powerful.

fade :: Fade

Card fade-in or fade-out

fadeOut :: [Tile]

Tiles to fade out atop the card spot, both when there's a creature and when there's not.

scoreChange :: Int

Score changed

type InPlaceEffects = Map Card InPlaceEffect

type family InPlaceType (p :: Phase) where ...

Equations

InPlaceType Core = CardsOnTable

InPlaceType UI = InPlaceEffects

bottomSpotOfTopVisual :: Card -> Card

Returns a bottom position, by taking a position that makes sense visually I.e. if you give this method [TopLeft], it'll correspond to the [TopLeft] bottom position that you SEE; even if positions make sense for the top part. This method takes care of translating correctly.

TODO smelc rename me to bottomSpot@

setPart :: T p -> Player -> PlayerPart p -> T p

toHoleyInPlace :: T Core -> [(Player, Card, Maybe (Creature Core))]

toInPlaceCreature :: T Core -> Player -> Card -> Maybe (Creature Core)

toHand :: T p -> Player -> InHandType p

toPart :: T p -> Player -> PlayerPart p

data DeathCause

Animations recorded upon death. If you are looking for the same thing in the more general case of the creature not dying, loot at fadeOut.

Constructors

DeathByFear
Creature was killed by fear

DeathByBreathIce
Creature killed by BreathIce attack

DeathByTerror
Creature was killed by terror

NoDeath
Creature was not killed

UsualDeath
Creature was killed for a reason we do not track precisely

data Deco

Constructors

Forest

initial

Arguments

:: RandomGen r

=> r

-> Teams (Team, [Card Core])
The initial decks

-> (r, T Core)
The shared model, with its RNG updated; and the initial board

The initial board, appropriately shuffled with a rng, and the starting decks of both teams.

newtype HandIndex

Constructors

HandIndex

Fields

unHandIndex :: Int

type family InHandType (p :: Phase) where ...

Equations

InHandType p = [HandElemType p]

data Kind

Data whose only purpose is to be lifted as a type, for disambiguating classes instances.

Constructors

Mana

lookupHand :: [a] -> Int -> Either Text a

lookupHandM :: MonadError Text m => [a] -> Int -> m a

class Mappable p a where

Functions to update content within T values

Methods

adjust :: Player -> Card -> (a -> a) -> T p -> T p

Maps over the data at the given Player and Card

adjustMany :: Player -> [Card] -> (a -> a) -> T p -> T p

Maps over the data at the given Player and Card

insert :: Player -> Card -> a -> T p -> T p

Sets the data at the given Player and Card

update :: Player -> Card -> (a -> Maybe a) -> T p -> T p

Updates the data at the given Player and Card. Updates if the function returns Just _ or removes the data if f returns Nothing.

mapHand :: Player -> (InHandType p -> InHandType p) -> T p -> T p

class PlayerKindIndexed (k :: Kind) p a where

Class to generically getmapset over the data of a player

Methods

getpk :: Player -> T p -> a

mappk :: (a -> a) -> Player -> T p -> T p

setpk :: Player -> a -> T p -> T p

data PlayerPart (p :: Phase)

Constructors

PlayerPart

Fields

deco :: Map Card (DecoType p)

inPlace :: InPlaceType p

Cards on the board

inHand :: InHandType p

Cards in hand

mana :: ManaType p

score :: ScoreType p

The score of this player

stack :: StackType p

discarded :: DiscardedType p

team :: TeamType p

type family StackType (p :: Phase) where ...

Equations

StackType Core = [ID]

StackType UI = Int

small :: Model -> Teams Team -> CreatureID -> [Item] -> Player -> Card -> T Core

A board with a single creature in place. Hands are empty. Handy for debugging for example.

toStack :: T p -> Player -> StackType p

setStack :: T p -> Player -> StackType p -> T p

setHand :: T p -> Player -> InHandType p -> T p

addToHand :: T p -> Player -> HandElemType p -> T p

empty :: Empty a b => a -> b

increaseScore :: p ~ Core => T p -> Player -> Nat -> T p

mapDiscarded :: Player -> (DiscardedType p -> DiscardedType p) -> T p -> T p

Map over the discarded field of the given player in the given board

toDiscarded :: T p -> Player -> DiscardedType p

setInPlace :: Player -> InPlaceType p -> T p -> T p

toInPlace :: T p -> Player -> InPlaceType p

line :: Line -> [Card]

All spots in a given line

data StackKind

Constructors

Handed
Button to display the hand (only used for the enemy, in Debug mode)

Stacked
Button to display the stack

Discarded
Button to display discarded cards

toScore :: Player -> T p -> ScoreType p

mapScore :: T p -> Player -> (ScoreType p -> ScoreType p) -> T p

neighbors :: Neighborhood -> Card -> [Card]

data Neighborhood

The various kinds of neighbors

Constructors

Cardinal
Neighbors to the left and the right

Diagonal
Neighbors in diagonals

All
Cardinal + diagnoal neighbors

data T (p :: Phase)

Things on the board (hand excluded)

Constructors

T

Fields

playerTop :: PlayerPart p

playerBottom :: PlayerPart p

toPlayerHoleyInPlace :: T Core -> Player -> [(Card, Maybe (Creature Core))]

toNeighbors :: T Core -> Player -> Card -> Neighborhood -> [(Card, Creature Core)]

The neighbors of the card at the given spot, for the given player, and for the given kind of neighbors.

toPlayerCardSpots :: T Core -> Player -> CardTargetKind -> [Card]

Returns the spots available for playing a card with the given target kind. For creature cards for example it returns empty spots. For item cards it returns occupied spots.

isDead :: DeathCause -> Bool

Whether this death cause represents a death

toData :: Player -> Teams a -> a

type family ManaType (p :: Phase) where ...

Equations

ManaType Core = Nat

ManaType UI = Int

switchLine :: Card -> Card

Given a frontline spot, the corresponding backline spot. Given a backline spot, the corresponding frontline spot.