Campaign

Description

This module deals with the succession of matches |

Documentation

newtype Journey

For every level, the opponent and the theme. Maps of this type are complete: the domain is the entire of Level

Constructors

Journey (Map Level (Team, Theme))

data Level

The levels: Level0 is the first level (no reward was given yet), Level1 is the second level; etc.

Constructors

Level0

Level1

data Outcome

The outcome of playing a single game

Constructors

Win
Player wins

Draw
Draw

Loss
Player loses

anywhere :: [Team]

Teams that can be fought against at any Level. See fixed for other possible fights.

augment :: [ID] -> Level -> Team -> [[ID]]

Given the current deck, all decks that are possible at the given level

fixed :: [(Team, Level)]

Fights that can be done only at specific levels

mkJourney :: Team -> Journey

mkJourney team returns a journey for when the player plays team. In the future it will likely be randomized.

nbRewards :: Level -> Nat

The number of rewards that have been obtained when playing the given level

succ :: Level -> Maybe Level

Given a level, its successor

unsafeJourney :: Level -> Team -> Journey

An incomplete journey, but fine for playing one game at the given level, against the given team.