Card

Documentation

data Team

Constructors

Evil

Human

Undead

ppTeam :: Team -> String

allTeams :: [Team]

data Skill

Constructors

Blow

BreathIce

Discipline

DrawCard

Fear
Creature causes fear

LongReach

Ranged

Source Nat
Creature creates mana at beginning of turn

Stupid4

Terror
Creature causes terror

Unique

data SkillCore

Constructors

Blow' Bool
Whether the skill is available (True) or used already (False)

BreathIce'

Discipline'

DrawCard' Bool
Whether the skill is available (True) or used already (False)

Fear' Bool
Creature causes fear, the Boolean indicates whether the skill is available (True) or used already (False)

LongReach'

Ranged'

Source' Nat Bool
Creature create mana at beginning of turn. Boolean indicates whether the skill is available (True) or used already (False)

Stupid4' Int
The turn, at 0, 1, or 2 not stupide; at 3 stupid; then back to 0

Terror' Bool
Creature causes terror, the Boolean indicates whether the skill is available (True) or used already (False)

Unique'

isStupid :: SkillCore -> Bool

data SkillPack

Constructors

SkillPack

Fields

skill :: Skill

skillText :: String

skillTitle :: String

data Phase

Constructors

Core
Phase for data in core algorithms (AI, Game); not for rendering. Creatures in this phase have actual hitpoints, actual attack, etc.

UI
Phase for data in UI algorithms: contains more data related to drawing cards. Data in this phase is _formal_, i.e. hitpoints are the maximum (pristine) hitpoints, attack is formal attack (before maluses or bonuses), etc.

type family ItemType (p :: Phase) where ...

Equations

ItemType UI = ItemObject UI

ItemType Core = Item

type family ManaType (p :: Phase) where ...

Equations

ManaType UI = Nat

ManaType Core = ()

type family OffsetType (p :: Phase) where ...

Equations

OffsetType UI = Int

OffsetType Core = ()

type family SkillType (p :: Phase) where ...

Equations

SkillType UI = Skill

SkillType Core = SkillCore

type family TeamsType (p :: Phase) where ...

Equations

TeamsType UI = [Team]

TeamsType Core = ()

type family TextType (p :: Phase) where ...

Equations

TextType UI = String

TextType Core = ()

type family TileType (p :: Phase) where ...

Equations

TileType UI = Tile

TileType Core = ()

type family TransientType (p :: Phase) where ...

Equations

TransientType UI = ()

TransientType Core = Bool

type Forall (c :: Type -> Constraint) (p :: Phase) = (c (ItemType p), c (ManaType p), c (OffsetType p), c (SkillType p), c (TeamsType p), c (TextType p), c (TileType p), c (TransientType p))

data CreatureKind

Constructors

Archer

Church

General

Ghost

Knight

Necromancer

Mummy

Skeleton

Shade

Spearman

Specter

Swordsman

Ogre

Vampire

Warrior

allCreatureKinds :: [CreatureKind]

data CreatureID

Constructors

CreatureID

Fields

creatureKind :: CreatureKind

team :: Team

isSkeleton :: CreatureID -> Bool

data Creature (p :: Phase)

Constructors

Creature

Fields

creatureId :: CreatureID

hp :: Nat

attack :: Nat

Beware when using this accessor, you may want totalAttack instead

items :: [ItemType p]

moral :: Int

skills :: [SkillType p]

transient :: TransientType p

Creature doesn't go to the discarded stack when killed

class Itemizable a where

Methods

getItems :: a -> [Item]

data Neutral

Constructors

Health

InfernalHaste

Life

Plague

allNeutrals :: [Neutral]

data NeutralObject (p :: Phase)

Constructors

NeutralObject

Fields

neutral :: Neutral

neutralTeams :: TeamsType p

The teams to which this neutral card applies

ntext :: TextType p

ntitle :: TextType p

data Item

Constructors

Crown

FlailOfTheDamned

SkBanner

SwordOfMight

allItems :: [Item]

data ItemObject (p :: Phase)

Constructors

ItemObject

Fields

item :: Item

teams :: TeamsType p

itext :: TextType p

itextSzOffset :: OffsetType p

ititle :: TextType p

ititleSzOffset :: OffsetType p

mkCoreItemObject :: Item -> ItemObject Core

data CardCommon (p :: Phase)

Data that is used by all three kind of cards

Constructors

CardCommon

Fields

mana :: ManaType p

tile :: TileType p

mkCoreCardCommon :: CardCommon Core

data Card (p :: Phase)

Constructors

CreatureCard (CardCommon p) (Creature p)

NeutralCard (CardCommon p) (NeutralObject p)

ItemCard (CardCommon p) (ItemObject p)

toCommon :: Card p -> CardCommon p

liftSkill :: SkillCore -> Skill

class Unlift t where

Methods

unlift :: t UI -> t Core

unliftSkill :: Skill -> SkillCore

Because this function uses default values, it is NOT harmless! Use only when initializing data.

cardToCreature :: Card p -> Maybe (Creature p)

cardToItemObject :: Card p -> Maybe (ItemObject p)

cardToNeutralObject :: Card p -> Maybe (NeutralObject p)

data ID

The minimal identifier of a card. See SharedModel to obtain | a full-fledged card from that.

Constructors

IDC CreatureID [Item]

IDI Item

IDN Neutral

cardToIdentifier :: Itemizable (Creature p) => Card p -> ID

identToId :: ID -> Maybe CreatureID

groupCards :: Itemizable (Creature p) => [Card p] -> Map ID [Card p]

teamDeck

Arguments

:: [Card UI]
The cards as loaded from disk

-> Team
The team for which to build the deck

-> [Card Core]
The initial deck

data CardTargetKind

Constructors

Hole
Card targets empty CardSpot

Occupied
Card targets occupied CardSpot

data TargetType

Constructors

CardTargetType CardTargetKind
Target is a single card

PlayerTargetType
Target is an entire part of the bard

targetType :: ID -> TargetType

The kind of Game target a neutral likes