CardInstances

Description

This module contains instances of classes regarding types defined in Card. |

Documentation

class Startable a where

Class for things that can be reset at the start of a turn

Minimal complete definition

Nothing

Methods

start :: a -> a