Cinema

Description

Module containing the base for animating scenes in lobbies

Documentation

data Actor

Constructors

Actor

Fields

actorName :: Maybe String

actorState :: ActorState

data ActorKind

Constructors

CreatureKind

TileKind

data ActorState

Constructors

ActorState

Fields

direction :: Direction

In which direction the sprite is looking at

sprite :: Maybe Sprite

How to draw this actor

telling :: Maybe String

Whether the element says something

x :: Int

0 means on the left

y :: Int

0 means top

data Direction

Constructors

ToLeft

ToRight

data Element

newtype Frame

Constructors

Frame

Fields

unFrame :: Map Element Actor

type Scene = Program SceneInstruction

data TimedFrame

Constructors

TimedFrame

Fields

duration :: Duration

The duration of a frame, in tenth of seconds

frame :: Frame

The frame

(|||) :: Scene () -> Scene () -> Scene ()

creatureSprite :: CreatureID -> Sprite

defaultDirection :: Direction

dress :: Sprite -> Element -> Scene ()

down :: Element -> Scene ()

dot :: Element -> (ActorState -> a) -> Scene a

Typical usage: archerX <- archer dot x

getActorState :: Element -> Scene ActorState

fork :: Scene () -> Scene ()

forkWithId :: Scene () -> Scene ThreadId

joinThread :: ThreadId -> Scene ()

hide :: Element -> Scene ()

left :: Element -> Scene ()

newActor :: Maybe String -> ActorState -> Scene Element

newActorAt :: String -> Sprite -> Int -> Int -> Scene Element

newActorAt' :: String -> Sprite -> Direction -> Int -> Int -> Scene Element

render :: Scene () -> [TimedFrame]

right :: Element -> Scene ()

setActorState :: Element -> ActorState -> Scene ()

shutup :: Element -> Scene ()

spriteToKind :: Sprite -> ActorKind

tell :: String -> Element -> Scene ()

tileSprite :: Tile -> Sprite

turnAround :: Element -> Scene ()

up :: Element -> Scene ()

during :: Duration -> Scene a -> Scene a

(=:) :: Element -> ActorState -> Scene ()

newHiddenActor :: String -> Scene Element

moveTo :: Int -> Int -> Element -> Scene ()

resetAt :: Sprite -> Int -> Int -> Element -> Scene ()

turnTo :: Direction -> Element -> Scene ()

resetAt' :: Sprite -> Direction -> Int -> Int -> Element -> Scene ()

wait :: Duration -> Scene ()