Command

Description

This module contains the debug commands that can be entered in dev mode. It is in the shared/ folder, for testing purpose: we don't want to compile client/ in tests. |

Documentation

data Command

Constructors

AIPlay Player
Make the AI play at a spot

Assassins Player
Command to trigger the assassins event

CreateForest Player
Command to trigger the 'create forest' event

EndGame Outcome
Command to end the current game. Parameter indicates how the playing player performed.

FillTheFrontline Player
Command to trgger the 'fill the frontline' event

Gimme ID
Command to obtain an extra card in the hand in GameView

GimmeMana
Command to obtain mana

Growth Player
Command to trigger the growth event

HailToTheKing Player
Command to trigger the king event

Killall Player
Command to kill all creatures in a part

Reboot Player Team
Command to reset the state of a part, as if it had started the game with the given team.

allCommands :: [CreatureID] -> [Command]