Configuration

Documentation

data Edition

Constructors

Vanilla

Legendary

data Configuration

Constructors

Dev
Dev version: https://www.schplaf.org/hgames/darkcraw/ This version is pushed regularly by @smelc. Second parameter is the commit hash.

Itch Edition String
Version released on https://hgames.itch.io/pixel-card-wars This version is pushed when relevant by @smelc (where there's a devlog with it usually). Second parameter is the commit hash.

Schplaf Edition String
Version released on https://www.schplaf.org/hgames/darkcraw This version is usually pushed at every commit by @smelc. Second parameter is the commit hash.

isDev :: Bool

get :: Configuration

The current configuration