Damage

Description

Module regarding damage, based on dice damage like in DnD

Documentation

data Damage

Constructors

Damage

Fields

base :: Nat

The base damage amount. Damage dealt is equal or greater than this value.

variance :: Nat

The variance added on top of base. If '0', then damage is fixed. If greater than 0, a dice is rolled to add from '0' (included) to variance (included) damage. So if you want '1-2' damage range, have 'base == 1' and 'variance == 1'.

(+^) :: Damage -> Nat -> Damage infixl 6

'a +^ n' augments the base damage of a by n

(-^) :: Damage -> Nat -> Damage infixl 6

'a -^ n' reduces the base damage of a by n (if possible)

class Dealer a where

Whether something can deal damage. This function is meant to be used qualified.

Methods

dealer :: a -> Bool

mean :: Damage -> Nat