Effect

Description

Module defining core and UI effects. Core effects affect boards while UI effects affect UI boards.

Documentation

data DeathCause

Animations recorded upon death. If you are looking for the same thing in the more general case of the creature not dying, look at fadeOut. TODO smelc remove me, superseded by fadeOut in T@.

Constructors

DeathByBleed
Creature was killed by bleeding

DeathByFear
Creature was killed by fear

DeathByBreathIce
Creature killed by BreathIce attack

DeathByTerror
Creature was killed by terror

NoDeath
Creature was not killed

UsualDeath
Creature was killed for a reason we do not track precisely

isDead :: DeathCause -> Bool

Whether this death cause represents a death

data DecoChange

A change in the value of the deco map

Constructors

NoDecoChange
No change

Appears Deco
A new Deco appears

data Deco

Constructors

Forest

data T

Initially this type was for displaying animations only. However Game also uses for Core stuff internally (see applyEffectOnBoard). Unfortunate :-( So be careful when changing related code.

This type mostly deals with the defender, even though there is the attackBump field (could this field be removed?)

Constructors

T

Fields

attackChange :: Int

Attack value changed

block :: Bool

Defender used the Block skll

decoChange :: DecoChange

Whether a Deco changes

death :: DeathCause

Did creature die? If yes, for what reason

attackBump :: Bool

Creature attacked

hitPointsChange :: Int

Hits points changed

extra :: Nat

Hit points that were not dealt because the defender died. Contributes to the score with Powerful.

fade :: Fade

Card fade-in or fade-out

fadeOut :: [Tile]

Tiles to fade out atop the card spot, both when there's a creature and when there's not.

scoreChange :: Int

Score changed

type Ts = Map Card T