Event

Documentation

onMouseEnter' :: Text -> Action -> Attribute Action

Like onMouseEnter, but restricts event firing to elements with the provided class name.

onMouseLeave' :: Text -> Action -> Attribute Action

Like onMouseLeave, but restricts event firing to elements with the provided class name.