Game

Documentation

allEnemySpots

Arguments

:: Card
The attacker's position

-> [Card]
Spots that can be attacked

The other spot in the column in the spot's part otherYSpot :: Spots.Card -> Spots.Card otherYSpot TopLeft = BottomLeft otherYSpot Top = Bottom otherYSpot TopRight = BottomRight otherYSpot BottomLeft = TopLeft otherYSpot Bottom = Top otherYSpot BottomRight = TopRight

Spots that can be attacked from a spot. Spot as argument is in one player part while spots returned are in the other player part. The order in the result matters, the first element is the first spot attacked, then the second element is attacked if the first spot is empty or if the creature can attack multiple spots for some reasons.

data Animation

An animation that makes sense at the Game level. If you consider extending this variant, consider whether it would be be better in Board 'UI

Constructors

Application Player Target (Card Core)
Player plays the given card played on a target. This is used for example to display Plague when played or when adding an item to a creature.

NoAnimation

Fadeout
Game view should fadeout

Message [MessageText] Nat
Message to show centered. The Nat is the duration during which to show the message. Then it fades out during one second.

appliesTo :: Board Core -> ID -> Player -> Target -> Bool

board id pSpot target holds iff player at pSpot can play card id on target

applyFearNTerror

Arguments

:: Board Core
The input board

-> Player
The part causing fear

-> (Board Core, Board UI)

applyFillTheFrontline :: Board Core -> Player -> Board Core

applyPlague

Arguments

:: Board Core

-> Player
The part on which to apply plague

-> Board Core

attackOrder :: Player -> [Card]

The order in which cards attack

cardsToDraw :: Board Core -> Player -> Bool -> [DrawSource]

The cards to draw, the Boolean indicates whether to bound by the stack's length or not

drawCards

Arguments

:: SharedModel

-> Board Core

-> Player
The player drawing cards

-> [DrawSource]
The sources from which to draw the cards

-> Either Text (SharedModel, Board Core, Board UI)

enemySpots

Arguments

:: Creature Core
The attacker

-> Card
Where the attack is

-> EnemySpots [Card]

Spots that can be attacked by a creature. Spot as argument is in one player part while spots returned are in the other player part. The order in the result matters, the first element is the first spot attacked, then the second element is attacked if the first spot is empty or if the creature can attack multiple spots for some reasons.

idToHandIndex :: Board Core -> Player -> ID -> Maybe HandIndex

The index of the card with this ID, in the hand of the player at the given spot

data DrawSource

The reason for drawing a card

Constructors

Native
Drawing one of the [nbDrawCards] cards allowed

CardDrawer Player Card
Drawing a card because of a creature with the [DrawCard] skill at the given position

eventToAnim :: SharedModel -> Board Core -> Event -> Animation

Translates an Event into an animation displayed in the middle of the Board.

keepEffectfull :: SharedModel -> Board Core -> [Event] -> [Event]

'keepEffectfull board es' returns the elements of es that have an effect. Elements of es are played in sequence.

data EnemySpots a

Type to handle Imprecise when resolving attacks

Constructors

Imprecise
Creature has Imprecise, enemy spots doesn't make sense

Spots a
A creature without Imprecise

data Event

Constructors

ApplyChurch Player
Apply church of the creatures at the given Player

ApplyFearNTerror Player
Apply fear caused by the creatures at the given Player

ApplyKing Player
Apply king of the creatures at the given Player

Attack Player Card Bool Bool
A card attacks at the given spot. The first Boolean indicates whether the next spot (as defined by nextAttackSpot) should be enqueued after solving this attack. The second Boolean indicates whether GameIncrTurn (change player turn) should be performed after solving this stream of attacks.

FillTheFrontline Player
Ranged creatures with the Ranged skill, that: 1/ are in the back line and 2/ have no creature in from of them; move frontward

NoPlayEvent
A Nothing case, for convenience

Place Player Target HandIndex
Player puts a card from his hand on its part of the board. First argument is the player, second argument is the target, third argument is the card being played.

Place' Player Target ID
AI puts a card from his hand. This constructor has better testing behavior than Place: it makes the generated events commute.

data MessageText

Constructors

Text Text
Simple text to display

Image Filepath
Constructor to display an image. The image should be fine for passing to assetsPath

data PolyResult a

The polymorphic version of Result. Used for implementors that do not return events and hence instantiate a by ()

Constructors

PolyResult SharedModel (Board Core) a (Board UI)

type Result = PolyResult (Maybe Event)

The result of playing an Event: an updated board, the next event (if any), and the animations

nextAttackSpot :: Board Core -> Player -> Maybe Card -> Maybe Card

nextAttackSpot b pSpot cSpot returns the spots to attack after cSpot. If cSpot is Nothing, the first spot in the order is considered; if cSpot is Just _, then spots after cSpot are considered.

play :: SharedModel -> Board Core -> Event -> Either Text Result

Play a single Event on the given Board

playAll :: SharedModel -> Board Core -> [Event] -> Either Text (PolyResult ())

Please avoid calling this function if you can. Its effect is difficult to predict because of cards that create events, as these events are played right away (instead of being enqueued in the main loop like when playing Infernal Haste interactively). FIXME @smelc take an additional parameter telling whether to mimic the main loop (for Game.Attack events).

playM :: MonadError Text m => MonadWriter (Board UI) m => MonadState SharedModel m => Board Core -> Event -> m (Board Core, Maybe Event)

data StatChange

A change in stats of a creature. Using Nat for now because it suffices for the use cases. But really it could be a Int, it just makes apply harder.

Constructors

StatChange

Fields

attackDiff :: Nat

hpDiff :: Nat

transferCards :: SharedModel -> Board Core -> Player -> (SharedModel, Board Core, Board UI)

data Target

On what a card can be applied

Constructors

PlayerTarget Player
Neutral card applies to all in place cards of a player

CardTarget Player Card
Creature card placed at given spot or Neutral card applies to a given in place card of a player

whichPlayerTarget :: ID -> WhichPlayerTarget

data WhichPlayerTarget

Whether a card makes sense on the playing player or the opponent. We could even try both, but we don't do that for now

Constructors

Playing

Opponent