GameView

Description

Module to display an instance of the main view, i.e. the battle between two players. Only this module (and modules that depend on it) can depend on GameViewInternal.

Documentation

viewGameModel :: Game -> Styled (View Action)

Constructs a virtual DOM from a game model

mkTargetOffset :: Target -> Map MisoString MisoString

mkTargetOffset returns the offset to display the application of a Neutral card to a creature in place.

cmdDiv :: Model -> [View Action]

data InPlaceCellContext

Dumb container to reduce the number of arguments to some functions of this file.

Constructors

InPlaceCellContext

Fields

z :: Int

The z-index

mkOffset :: Player -> Card -> Map MisoString MisoString

How to build the style for offsetting a card from the board's origin

boardToInPlaceCells

Arguments

:: InPlaceCellContext

-> Game

-> Maybe TargetType
The target to which the card being dragged applies (if any)

-> Styled [View Action]

boardToInPlaceCell

Arguments

:: InPlaceCellContext

-> Game

-> InPlaceActionizer (Player, Card) Action

-> Maybe TargetType
The target to which the card being dragged applies (if any)

-> Player
The part considered

-> Card
The spot of the card to show

-> Styled (View Action)

decoViews :: Int -> Player -> T Core -> [View Action]

boardToPlayerTarget :: Int -> Game -> Maybe TargetType -> Player -> Maybe (View Action)

Whether to show the player target at pSpot

dragDropEvents :: Target -> [Attribute Action]

The events for placeholders showing drag targets

borderRGB :: Eq a => Interaction a -> a -> (Int, Int, Int)

boardToInHandCells

Arguments

:: Int
The z index

-> HandDrawingInput

-> Styled [View Action]

toHandDrawingInput :: Game -> HandDrawingInput

data InPlaceActionizer a b

Events on in place cards: hovering

Constructors

InPlaceActionizer

Fields

onMouseEnter :: a -> b

onMouseLeave :: a -> b

data HandActionizer a b

Events on cards in hand

Constructors

HandActionizer

Fields

inPlaceActionizer :: InPlaceActionizer a b

onDragStart :: a -> b

onDragEnd :: b

type HandOffseter = (Int, Int) -> (Int, Int)

data HandDrawingInput

Constructors

HandDrawingInput

Fields

hand :: [(Card Core, Bool)]

The hand, and whether the corresponding card is being faded in

interaction :: Maybe (Interaction Target)

The current interaction, if any. FIXME @smelc do not hardcode Game.Target

labeler :: Mana -> String

How to typeset mana

mana :: Nat

The mana of the team being drawn

offseter :: HandOffseter

How to offset the default position of the hand's cards

team :: Team

The team of the hand being drawn

turn :: T

The current turn

playingPlayer :: Player

The player of the hand being drawn

shared :: Model

boardToInHandCell :: HandDrawingInput -> HandActionizer HandIndex Action -> Int -> (Int, Card Core, HandIndex) -> Styled (View Action)

cardCellsBoardOffset :: Player -> Card -> (Int, Int)

Offsets, in number of cells, from the board's origin

handCell :: View Action