GameViewInternal

Description

This module contains things used in GameView.hs but which are likely rarely modified, to avoid GameView.hs growing too much. Other modules besides GameView should nod depend on this module. |

Documentation

animToFade :: Animation -> Fade

What fading a Animation triggers

newtype Border

Data for drawing a border. The only data is the color, since for now all borders have the same width.

Constructors

Border

Fields

unBorder :: T

data Borders

Borders of all boxes

Constructors

Borders

Fields

applications :: [Box]

Boxes where something can be applied

borders :: Map Box Border

Box coming from hovering

errView :: Game -> Int -> Maybe (View Action)

fadeouts :: Model -> Int -> T -> Styled (Maybe (View Action))

heartWobble :: Int -> T -> Styled [View Action]

hintView :: Int -> Game -> Borders -> Maybe (View Action)

keyframes :: MisoString -> MisoString -> [(Int, String)] -> MisoString -> MisoString

manaView :: Game -> Int -> View a

messageView :: Animation -> Maybe (Styled (View a))

mkBorders :: Game -> Borders

noDrag :: Attribute a

scoreViews :: Game -> Int -> [Styled (View Action)]

data StackPosition

Constructors

Board
Stack/discarded widget in board (enemy), only in hashless Configuration

Hand
Stack/discarded widget in hand (playing player) | Always visible

stackView :: Game -> Int -> Player -> StackPosition -> StackKind -> Styled [View Action]

The widget showing the number of cards in the stack/discarded stack

turnView :: Game -> Int -> Styled (View Action)

statChange :: Int -> StatChangeKind -> T -> Styled [View Action]

data StatChangeKind

Constructors

HitPoint

Attack