LootView

Documentation

view :: LootModel -> Styled (View Action)

Module to display the view reached after a successful end of GameView. This view sits between two games, where the player can pick new cards to augment its deck. |

Top-level rendering function

data Context

Context that is common to all calls to deckView

Constructors

Context

Fields

shared :: SharedModel

team :: Team

rewardsViewTopMargin :: Int

rewardsViewHeight :: Int

rewardsView :: Context -> Int -> [(ID, Picked)] -> Styled (View Action)

The div showing the rewards. It's displayed above deckView

deckViewTopMargin :: Int

deckViewHeight :: Int

deckCardsPerRow :: Int

deckViewWidth :: Int

deckView :: Context -> Int -> [ID] -> Styled (View Action)

The div showing the deck