MCTSAI

Documentation

place :: Difficulty -> Model -> Place -> Board Core -> Maybe (Board Core)

Plays one event and plays following events if any. The point of this function is to get the next state after playing one card. If the card cannot be played, Nothing is returned. We cannot use maybePlay because it doesn't enqueue the created events. We cannot use playAll because it doesn't distinguish if the first event succeeded. If it didn't, we MUST know it, so that the AI skips this event.

newPlaySim

Arguments

:: Difficulty

-> Model

-> Player
The spot to play

-> Board Core
The board to consider

-> [Place]

Generated events, after having chosen from the ones given by newPlay, by running the simulation.