Mana

Description

Module dealing with mana

Documentation

amount :: Turn -> Mana -> Nat

amount t m returns the amount of mana of m at turn t

class Labeler a where

How to typeset mana when an a is available

Methods

labeler :: a -> Mana -> String

data Mana

Different mana costs

Constructors

Const Nat
Mana cost is constant

RemainingNbOfTurns
Mana cost equals the number of remaining turns

class Read a where

A safe alternative to prelude's Read

Methods

read :: String -> Maybe a

(>=) :: Turn -> Mana -> Nat -> Bool

>= t m n returns whether m >= n when it's turn t

(<=) :: Turn -> Mana -> Nat -> Bool

<= t m n returns whether m <= n when it's turn t

(>) :: Turn -> Mana -> Nat -> Bool

> t m n returns whether m > n when it's turn t

(<) :: Turn -> Mana -> Nat -> Bool

< t m n returns whether m < n when it's turn t

show :: Maybe Turn -> Mana -> String

The text to show for mana, depending on whether the turn is provided or not