Mechanics

Description

Spin-off Game, to avoid Game to be gigantic

Documentation

data Move

Constructors

Move

Fields

from :: (Player, Card)

to :: (Player, Card)

mkEndoMove :: Player -> Card -> Card -> Move

A move within one Player

flySpot :: MonadState Model m => InPlaceType Core -> m (Maybe Card)

Whether a creature with Flying in the given place can fly to.

isRanged :: Creature Core -> Bool

Some definition of ranged. Use for fill the frontline for example, and for not being blocked by an ally.

move :: Move -> T Core -> T Core

move m board moves the creature at move.from m to move.to m Returns board unchanged if the move is impossible