Model

Documentation

data Interaction a

An interaction happening in the game page

Constructors

HoverInteraction Hovering
Hovering over a card in hand

HoverInPlaceInteraction a
Hovering over a target

DragInteraction (Dragging a)
Dragging a card

NoInteraction

ShowErrorInteraction Text

newtype Hovering

Constructors

Hovering

Fields

hoveredCard :: HandIndex

data Dragging a

Constructors

Dragging

Fields

draggedCard :: HandIndex

dragTarget :: Maybe a

data HandFiddle

Constructors

HandHovering HandIndex
Card in hand being hovered

HandDragging HandIndex
Card in hand being dragged

data Game

The model of the gaming page. If you add a field, think about extending the Show instance below.

Constructors

Game

Fields

shared :: SharedModel

Part of the model shared among all pages

board :: Board Core

The core part of the model

difficulty :: Difficulty

The game's difficulty

interaction :: Interaction Target

What user interaction is going on

journey :: Journey

The list of games to play

level :: Level

The current level

playingPlayer :: Player

Where the player plays

playingPlayerDeck :: [ID]

The deck of $sel:playingPlayer:Game

turn :: Turn

The current turn

uiAvail :: Bool

Whether interactions are possible right now

anims :: Board UI

Animations to perform next

anim :: Animation

Animation unrelated to Board

gameToDeck :: Game -> [ID]

endGame :: Game -> Outcome -> LootModel

unsafeLootModel :: WelcomeModel -> Model

Function for debugging only. Used to make the game start directly on the LootView.

unsafeGameModel :: WelcomeModel -> Model

Function for debugging only. Used to make the game start directly on the GameView. Similar to a function in Update (but we don't want to grow Update if we can avoid).

isPlayerTurn :: Game -> Bool

Whether it's the turn of the playing player, i.e. neither the AI turn | nor the turn of the other player if in multiplayer.

data PlayingMode

Constructors

NoPlayingMode

MultiPlayer

SinglePlayer

SinglePlayerTeam Team

type TimedFrames = Vector TimedFrame

data SceneModel

Constructors

SceneNotStarted TimedFrames

ScenePlaying TimedFrames Int

ScenePausedForDebugging TimedFrames Int

SceneComplete TimedFrames

data WelcomeModel

The model of the welcome page

Constructors

WelcomeModel

Fields

sceneModel :: SceneModel

shared :: SharedModel

Part of the model shared among all pages

keysDown :: Set Int

Keys currently down

data SinglePlayerLobbyModel

Constructors

SinglePlayerLobbyModel

Fields

singlePlayerLobbyTeam :: Maybe Team

The chosen team

singlePlayerLobbyShared :: SharedModel

Part of the model shared among all pages

data MultiPlayerLobbyError

Constructors

InvitationCancelledError UserName

InvitationRejectedError UserName

UserBusyError UserName

data MultiPlayerLobbyModel

Constructors

CollectingUserName UserName

WaitingForNameSubmission UserName

DisplayingUserList (Maybe MultiPlayerLobbyError) UserName [UserName]

InvitingUser UserName [UserName] UserName InvitationActorState

Invited UserName [UserName] UserName InvitedActorState

GameStarted UserName UserName

data InvitationActorState

Constructors

WaitingForUserInvitationAck

WaitingForRSVP

WaitingForInvitationDropAck

data InvitedActorState

Constructors

CollectingUserRSVP

WaitingForRejectionAck

WaitingForAcceptanceAck

data Deck

Constructors

Deck

Fields

deck :: [Card Core]

The deck to show

deckBack :: Model

The model to use when closing the deck view

player :: Player

To which player $sel:deckBack:Deck belongs

team :: Team

To which team the deck being shown belongs

shared :: SharedModel

Part of the model shared among all pages

data Picked

Whether a card has been picked or not in the LootView

Constructors

Picked
Reward has been chosen

NotPicked
Reward has not been chosen

data LootModel

Constructors

LootModel

Fields

nbRewards :: Nat

The number of rewards to be picked from $sel:rewards:LootModel

next :: Level

The next level

deck :: [ID]

The deck of the playing player

team :: Team

To which team the deck being shown belongs

shared :: SharedModel

Part of the model shared among all pages

rewards :: [(ID, Picked)]

The possible rewards, and whether they have been picked already or not.

data Model

The top level model

Constructors

DeckModel' Deck

GameModel' Game

LootModel' LootModel

SinglePlayerLobbyModel' SinglePlayerLobbyModel

WelcomeModel' WelcomeModel

MultiPlayerLobbyModel' MultiPlayerLobbyModel