Network

Description

Module providing the API using the data from Roads Also contains some hardcoded data, which - contrary to Roads - is not generated.

Documentation

chooseTeamSpots :: Map Team Coord

The position where to choose the team

data Encounter

Possible encounters on the world map

Constructors

Fight Team Kind
Fighting

Reward Nat
Picking up rewards. The Nat indicates the number of rewards.

Select Team
Choosing your team at the start of the game

fightSpots :: Map Team [(Coord, Kind)]

The position of fights, hardcoded yes

journeys :: Map Team [[Encounter]]

Journeys of playable teams, for testing. This data could be generated automatically, this is tracked in https://github.com/smelc/miso-darkcraw/issues/13

lootSpots :: Map Coord Nat

Map from coordinates to the number of loots at this position

class Network b where

The high-level API to query data from Roads

Methods

neighbors :: b -> Coord -> [Coord]

mkTopology :: [(Nat, Nat)] -> Topology

How to obtain the API from data in Roads

rewards :: Map Team [ID]

data Topology