PCWViewInternal

Description

This module contains generic things to be used in *View.hs and *ViewInternal.hs files. Contrary to ViewInternal, it contains things that are specific to the game; not solely pure HTML/CSS stuff. |

Documentation

cardBoxShadowStyle

Arguments

:: (Int, Int, Int)
The (r, g, b) of the border

-> Int
The width of the border

-> MisoString
The timing-function of the transition

-> Attribute a

cardView

Arguments

:: DisplayLocation

-> Int
The z index

-> SharedModel

-> Team
The team of the card, for selecting the background

-> Card Core

-> CardDrawStyle

-> Styled (View Action)

Div displaying a card. Draws the picture and the background and delegate most stuff to cardView'

cardPositionStyle

Arguments

:: Int
The horizontal offset from the enclosing container, in number of cells

-> Int
The vertical offset from the enclosing container, in number of cells

-> Map MisoString MisoString

cardPositionStyle'

Arguments

:: Int
The horizontal offset from the enclosing container, in pixels

-> Int
The vertical offset from the enclosing container, in pixels

-> Map MisoString MisoString

scrollbarStyle :: Map MisoString MisoString

viewFrame :: DisplayMode -> Int -> SharedModel -> Frame -> View a

data CardDrawStyle

Constructors

CardDrawStyle

Fields

fadeIn :: Bool

Whether the card should fade in

hover :: Bool

Whether the card is being hovered

data DisplayLocation

Where a card is being drawn

Constructors

GameInPlaceLoc Part

GameHandLoc

GameDragLoc

DeckLoc

LootLoc

data DisplayMode

Constructors

NormalMode

DebugMode

Orphan instances

ToMisoString Nat

Instance details

Methods

toMisoString :: Nat -> MisoString