Shared

Description

This module contains the subset of the model that is needed by Game I didn't want to make Game depend on Model, because it seemed overkill and dangerous cyclewise in the long run. Hence I introduced this module. |

Documentation

creatureToFilepath :: Model -> Creature UI -> Maybe Filepath

Maps a creature to the filepath of its tile.

identToCard :: Model -> ID -> Maybe (Card UI)

idToCreature :: Model -> CreatureID -> [Item] -> Maybe (Creature UI)

class Lift p where

Methods

lift :: Model -> p Core -> p UI

class Mlift p where

Methods

mlift :: Model -> p Core -> Maybe (p UI)

data Model

The part of the model that is likely to be used by all pages i.e. all possible models ! If you add a field, extend the EQ instance below. ! Note that we MUST have an EQ instance, because ultimately, our top-level model should have one. It is required by miso in the startApp function. It makes sense, since miso does diffs on models.

tileToFilepath :: Model -> Tile -> Size -> Filepath

unsafeGet :: Model

An instance of Model that is fine for debugging. Don't use it in production!

unsafeIdentToCard :: Model -> ID -> Card UI

liftSkill :: Model -> Skill -> Pack

create :: [Card UI] -> [Pack] -> [TileUI] -> StdGen -> Model

getStdGen :: Model -> StdGen

getCards :: Model -> [Card UI]

withCmd :: Model -> Maybe String -> Model

withStdGen :: Model -> StdGen -> Model

getCardIdentifiers :: Model -> [ID]

cardToFilepath :: Model -> Card UI -> Filepath

withSeed :: Model -> Int -> Model

allCommands :: Model -> [Command]

toCardCommon :: Model -> ID -> Maybe (CardCommon UI)

unsafeToCardCommon :: Model -> ID -> CardCommon UI

identToItem :: Model -> Item -> ItemObject UI

identToNeutral :: Model -> Neutral -> NeutralObject UI

getInitialDeck :: Model -> Team -> [Card Core]

The starting deck of the given team, not shuffled

Orphan instances

Eq StdGen

Instance details

Methods

(==) :: StdGen -> StdGen -> Bool

(/=) :: StdGen -> StdGen -> Bool