SharedModel

Description

This module contains the subset of the model that is needed by Game I didn't want to make Game depend on Model, because it seemed overkill and dangerous cyclewise in the long run. Hence I introduced this module. |

Documentation

creatureToFilepath :: SharedModel -> Creature UI -> Maybe Filepath

Maps a creature to the filepath of its tile.

identToCard :: SharedModel -> ID -> Maybe (Card UI)

idToCreature :: SharedModel -> CreatureID -> [Item] -> Maybe (Creature UI)

class Lift p where

Methods

lift :: SharedModel -> p Core -> p UI

class Mlift p where

Methods

mlift :: SharedModel -> p Core -> Maybe (p UI)

data SharedModel

The part of the model that is likely to be used by all pages | i.e. all possible models

tileToFilepath :: SharedModel -> Tile -> Size -> Filepath

unsafeGet :: SharedModel

An instance of SharedModel that is fine for debugging. Don't use it in production!

unsafeIdentToCard :: SharedModel -> ID -> Card UI

liftSkill :: SharedModel -> Skill -> SkillPack

shuffle :: SharedModel -> [a] -> (SharedModel, [a])

Shuffles the second argument with the random generator of SharedModel. Returns the shuffle and the updated SharedModel

create :: [Card UI] -> [SkillPack] -> [TileUI] -> StdGen -> SharedModel

getStdGen :: SharedModel -> StdGen

getCards :: SharedModel -> [Card UI]

getCmd :: SharedModel -> Maybe String

withCmd :: SharedModel -> Maybe String -> SharedModel

withStdGen :: SharedModel -> StdGen -> SharedModel

getCardIdentifiers :: SharedModel -> [ID]

cardToFilepath :: SharedModel -> Card UI -> Filepath

withSeed :: SharedModel -> Int -> SharedModel

allCommands :: SharedModel -> [Command]

toCardCommon :: SharedModel -> ID -> Maybe (CardCommon UI)

unsafeToCardCommon :: SharedModel -> ID -> CardCommon UI

identToItem :: SharedModel -> Item -> ItemObject UI

identToNeutral :: SharedModel -> Neutral -> NeutralObject UI

pick :: SharedModel -> [a] -> (SharedModel, Maybe a)

Pick one element at random from the second argument, using the random generator of SharedModel. Returns the updated SharedModel and the picked element (being Just if the list non-empty)

getInitialDeck :: SharedModel -> Team -> [Card Core]

The starting deck of the given team, not shuffled

oneof

Arguments

:: MonadState SharedModel m

=> NonEmpty a
The elements from which something must be picked

-> m a

Returns a random element from the list, using sharedStdGen as the random generator.

Orphan instances

Eq StdGen

Instance details

Methods

(==) :: StdGen -> StdGen -> Bool

(/=) :: StdGen -> StdGen -> Bool