Skill

Description

A spinoff of Card |

Documentation

data T blow drawCard fear source stupid terror

The type of skills. Highly polymorphic, because we use two different instances. Callers should only use the concrete instances Skill and State.

Constructors

Blow blow

BreathIce

Charge

Discipline

DrawCard drawCard

Fear fear
Creature causes fear

King

Knight
Identifier of knight

LongReach

Ranged

Source source
Creature creates mana at beginning of turn

Squire
Knight in front (if any) gains +1 hp and +1 attack upon arrival

Stupid4 stupid

Terror terror
Creature causes terror

Unique

Veteran
Upon arrival, every knight gains +1 hp and +1 attack

Zealot
Immune to fear

type Skill = T () () () Nat () ()

type State = T Bool Bool Bool (Nat, Bool) Nat Bool

isStupid :: State -> Bool

data Pack

Constructors

Pack

Fields

skill :: Skill

text :: String

title :: String

lift :: State -> Skill

unlift :: Skill -> State

Because this function uses default values, it is NOT harmless! Use only when initializing data.