Spots

Description

Module defining the player spots as well as the card spots. Historically this was in Board but was deported to a new file to be visible both from Board and Command. |

Documentation

data Card

The spot of a card, as visible from the top of the screen. For the | bottom part, think as if it was in the top, turning the board | 180 degrees clockwise; or use these values and map [bottomSpotOfTopVisual].

Constructors

TopLeft

Top

TopRight

BottomLeft

Bottom

BottomRight

data Player

The position of a player

Constructors

PlayerBot

PlayerTop

data Neighborhood

The various kinds of neighbors

Constructors

Cardinal
Neighbors to the left and the right

Diagonal
Neighbors in diagonals

All
Cardinal + diagnoal neighbors

Functions about CardSpot

allCards :: [Card]

All Card values.

allCardsNE :: NonEmpty Card

All Card values, as a non-empty list.

bottomSpotOfTopVisual :: Card -> Card

Returns a bottom position, by taking a position that makes sense visually I.e. if you give this method [TopLeft], it'll correspond to the [TopLeft] bottom position that you SEE; even if positions make sense for the top part. This method takes care of translating correctly.

frontSpots :: [Card]

Spots on the frontline

backSpots :: [Card]

Spots in the back line

inTheBack :: Card -> Bool

Whether a spot is in the back line

inFront :: Card -> Bool

Whether a spot is in the front line

neighbors :: Neighborhood -> Card -> [Card]

switchLine :: Card -> Card

Given a frontline spot, the corresponding backline spot. Given a backline spot, the corresponding frontline spot.

line :: Card -> [Card]

All spots in the same line

Functions about PlayerSpot

allPlayers :: [Player]

All Player values.

startingPlayerSpot :: Player

The spot of the player that starts a game. FIXME @smelc rename to starting and use me qualified.

endingPlayerSpot :: Player

The spot of the player that ends a game

other :: Player -> Player

The other Player spot. other . other == id