Total

Description

This module defines getters for getting the complete stat of a creature, boosts included. Functions of this module are meant to be called qualified. |

Documentation

affectedByFear :: Bool -> Creature Core -> Bool

Whether a creature is affected by fear. The first Boolean indicates whether to consider the hitpoints (True) or not (False)

affectedByTerror :: Bool -> Creature Core -> Bool

Whether a creature is affected by terror. The first Boolean indicates whether to consider the hitpoints (True) or not (False)

data Place

Where a creature is on the board

Constructors

Place

Fields

place :: InPlaceType Core

cardSpot :: CardSpot

mkPlace :: Board Core -> PlayerSpot -> CardSpot -> Place

Builds a Place from a Board

attack :: Maybe Place -> Creature Core -> Damage

The total attack of a creature, including boosts of skills and items. The Place indicates where the creature given is.

causesFear :: Creature Core -> Bool

Whether a creature causes fear

causesTerror :: Creature Core -> Bool

hasDiscipline :: [Skill] -> [Item] -> Bool

Core function for finding out about discipline

isDisciplined :: Creature Core -> Bool

Whether a creature has discipline