ViewBlocks

Description

This module defines the basic building blocks used by views, so that they have a uniform looknfeel |

Documentation

data ButtonState

Constructors

Disabled
Button is disabled

Enabled
Button is enabled

Selected
Button was enabled and has been chosen

dummyOn

Arguments

:: MisoString
One of "dragenter" or "dragover"

-> Attribute Action

Dummy onWithOptions instance. See https://github.com/dmjio/miso/issues/478

gui :: GUI a

data GUI a

Constructors

GUI

Fields

anyButton :: Int -> ButtonState -> ([Attribute a] -> View a) -> Styled (View a)

textButton :: Int -> ButtonState -> [Attribute a] -> MisoString -> Styled (View a)

argument #1 is the z-index | argument #2 is the button's state | argument #3 is the style | argument #4 is the button's text