WelcomeView

Description

Module to display the initial view of the game, i.e. the one that shows up when the game starts. |

Documentation

viewWelcomeModel :: Welcome -> Styled (View Action)

Constructs a virtual DOM from a welcome model