WorldView

Description

Module displaying the world map, which is the first view in legendary edition as well as the view between levels.

Documentation

viewWorldModel :: World -> Styled (View Action)

worldViewPixelHeight :: Int