Campaign

Description

This module deals with the succession of matches |

Documentation

data Level

The levels: Level0 is the first level (no reward was given yet), Level1 is the second level; etc.

Constructors

Level0

Level1

data Outcome

The outcome of playing a single game

Constructors

Win
Player wins

Draw
Draw

Loss
Player loses

decks :: [ID] -> Level -> Team -> [[ID]]

Given the initial deck, all decks that are possible at the given level

loot :: Outcome -> Level -> Team -> [ID]

The possible rewards when finishing the given Level with the given Outcome

nbRewards :: Level -> Nat

The number of rewards that have been obtained when playing the given level

succ :: Level -> Maybe Level

Given a level, its successor