Command

Description

This module contains the debug commands that can be entered in dev mode. It isin the shared/ folder, for testing purpose: we don't want to compile client/ in tests. |

Documentation

data Command

Constructors

EndGame Outcome
Command to end the current game. Parameter indicates how the playing player performed.

FillTheFrontline PlayerSpot
Command to trgger the 'fill the frontline' event

Gimme ID
Command to obtain an extra card in the hand in GameView

GimmeMana
Command to obtain mana

class Read a where

Methods

read :: String -> Maybe a

allCommands :: [CreatureID] -> [Command]

data PlayerSpot

We don't reuse Board type, because that would create a cyclic dependency

Constructors

Bot

Top