Command

Description

This module contains the debug commands that can be entered in dev mode. It is in the shared/ folder, for testing purpose: we don't want to compile client/ in tests. |

Documentation

data Command

Constructors

EndGame Outcome
Command to end the current game. Parameter indicates how the playing player performed.

FillTheFrontline PlayerSpot
Command to trgger the 'fill the frontline' event

Gimme ID
Command to obtain an extra card in the hand in GameView

GimmeMana
Command to obtain mana

HailToTheKing PlayerSpot
Command to trigger the king event

Killall PlayerSpot
Command to kill all creatures in a part

class Read a where

Methods

read :: String -> Maybe a

allCommands :: [CreatureID] -> [Command]

data PlayerSpot

The position of a player TODO @smelc rename me to Player and use me qualified

Constructors

PlayerBot

PlayerTop