Constants

Documentation

assetsDir :: MisoString

assetFilenameBeigeBG :: MisoString

assetFilenameCrown :: MisoString

assetFilenameGhost :: MisoString

assetFilenameHeart :: MisoString

assetFilenameMana :: MisoString

assetFilenameShade :: MisoString

assetFilenameSnowflake :: MisoString

assetFilenameSkull :: MisoString

assetFilenameSword :: MisoString

assetRoundTreeForestSpell :: MisoString

assetSlotBlue :: MisoString

assetSlotRed :: MisoString

assetsPath :: MisoString -> MisoString

assetsUrl :: MisoString -> MisoString

boardCellHeight :: Int

The number of vertical cells in a board

boardCellWidth :: Int

The number of horizontal cells in a board

boardToLeftCardCellsOffset :: Int

The number of cells from the left of the board to leftmost cards

boardPixelWidth :: Int

The board's width, in pixels

boardPixelHeight :: Int

The board's height , in pixels

blowAmount :: Nat

The additional attack granted by the Blow skill when available

borderSize :: Int

The size of borders around cards

buildPixelHeight :: Int

The height of the hand in the build view, in pixels

chargeAmount :: Nat

The additional attack granted by the Charge skill when available

nbCardsToDraw :: Int

The number of cards to draw at the beginning of a turn

data Difficulty

The AI's level. Not truly a constant, I know.

Constructors

Easy

Medium

Hard

data Fade

Fade or not fade. Not truly a constant. I know.

Constructors

FadeIn
Fadein (transparent->visible)

FadeOut
Fadeout (visible->transparent)

DontFade
Don't do any fadeing

gameTitle :: MisoString

handPixelHeight :: Int

The hands's height, in pixels

handPixelWidth :: Int

The hands's width, in pixels

initialHandSize :: Int

The initial number of cards in the hand

initialMana :: Nat

The initial mana of a team

nbTurns :: Nat

The number of turns during a game

squireAttackBonus :: Nat

The attack boost given by a squire to the knight in front

strengthPotAttackBonus :: Nat

The bonus given by a strength potion

cardBackground :: MisoString -> MisoString

cardCellHeight :: Int

A card's height, in cells

cardCellWidth :: Int

A card's width, in cells

cardPixelHeight :: Int

A card's height, in pixels

cardPixelWidth :: Int

A card's width, in pixels, equal to 24*3=72

cardHCellGap :: Int

The horizontal spacing between two cards, in cells

cardVCellGap :: Int

The vertical spacing between two cards, in cells

defaultManaCost :: Nat

The default mana prize if omitted in data.json

itchURL :: MisoString

The URL of the game on itch.io

teamsVCellGap :: Int

The vertical spacing between the two teams, in cells

turnPixelHeight :: Int

The turn widget height, in pixels

turnPixelWidth :: Int

The turn widget width, in pixels

cellPixelSize :: Int

The size of a cell, in pixels

cps :: Int

The size of a cell, in pixels

defaultFontSize :: Int

The default font size in pixels. It's actually the same as the default of browsers. | See e.g. https://www.w3schools.com/css/css_font_size.asp

seize :: Int

The size of smaller individual assets (items, heart, etc.)

subtitleFontSize :: Int

The size of subtitles

titleFontSize :: Int

The size of titles

lobbiesCellWidth :: Int

The width of the backgrounds of non-board view, in cells

lobbiesPixelWidth :: Int

The width of the backgrounds of non-board view, in pixels

lobbiesPixelHeight :: Int

The height of the backgrounds of non-board view, in pixels

manaLeftPixelOffset :: Int

The number of pixels from the left of the board, to the center of the mana column

beigeHTML :: MisoString

greenHTML :: MisoString

greyHTML :: MisoString

greenRGB :: (Int, Int, Int)

redHTML :: MisoString

yellowHTML :: MisoString

yellowRGB :: (Int, Int, Int)

whiteHTML :: MisoString