GameViewInternal

Description

This module contains things used in GameView.hs but which are likely rarely modified, to avoid GameView.hs growing too much. Other modules besides GameView should nod depend on this module. |

Documentation

borderWidth :: GameModel -> Target -> Int

Draw border around target if: 1/ card in hand is being hovered or dragged -> draw borders around valid drag targets or 2/ card in place is being hovered -> draw borders around cards be attacked from this card== playingPlayerSpot,

deathFadeout :: InPlaceEffect -> Int -> Int -> Styled [View Action]

errView :: GameModel -> Int -> Maybe (View Action)

heartWobble :: Int -> InPlaceEffect -> Styled [View Action]

keyframes :: MisoString -> MisoString -> [(Int, String)] -> MisoString -> MisoString

manaView :: GameModel -> Int -> View a

messageView :: Animation -> Maybe (Styled (View a))

noDrag :: Attribute a

scoreViews :: GameModel -> Int -> [View Action]

data StackPosition

Constructors

Board
Stack/discarded widget in board (enemy), only in hashless Configuration

Hand
Stack/discarded widget in hand (playing player) | Always visible

stackView :: GameModel -> Int -> PlayerSpot -> StackPosition -> StackKind -> Styled [View Action]

The widget showing the number of cards in the stack/discarded stack

turnView :: GameModel -> Int -> Styled (View Action)

statChange :: Int -> StatChangeKind -> InPlaceEffect -> Styled [View Action]

data StatChangeKind

Constructors

HitPoint

Attack