LootView

Description

Module to display the view reached after a successful end of GameView. This view sits between two games, where the player can pick new cards to augment its deck.

Documentation

view :: LootModel -> Styled (View Action)

Top-level rendering function

data Context

Context that is common to all calls to deckView

Constructors

Context

Fields

remainingToPick :: Nat

shared :: SharedModel

team :: Team

rewardsViewTopMargin :: Int

The margin from the top to the row of the rewards

rewardsViewHeight :: Int

The height of the rewards div: the height of a card plus space for the individual legends atop each card

rewardsView :: Context -> Int -> [(ID, Picked)] -> Styled (View Action)

The div showing the rewards. It is displayed above deckView

nextView :: Int -> Nat -> Styled (View action)

deckViewTopMargin :: Int

deckViewHeight :: Int

deckCardsPerRow :: Int

deckViewWidth :: Int

counts :: forall a. Eq a => [a] -> [(a, Nat)]

Gather equal items, counting them on the way; and *keep the order*. We could use Data.Map.Ordered, but it doesn't have mapFromListWith so it's not really easier.

data DeckCard a

A dumb container for passing structured data to deckView

Constructors

DeckCard

Fields

card :: a

The card

picked :: Maybe Nat

If the card comes from the reward, its index in the list of rewards otherwise Nothing.

deckView :: Context -> Int -> [DeckCard ID] -> Styled (View Action)

The div showing the deck. The value of type Picked indicates whether the card is a reward that was picked or a card from the deck of the previous game.