Skill

Description

A spinoff of Card

Documentation

data Skill

Constructors

Blow

BreathIce

Discipline

DrawCard

Fear
Creature causes fear

LongReach

Ranged

Source Nat
Creature creates mana at beginning of turn

Stupid4

Terror
Creature causes terror

Unique

data SkillCore

Constructors

Blow' Bool
Whether the skill is available (True) or used already (False)

BreathIce'

Discipline'

DrawCard' Bool
Whether the skill is available (True) or used already (False)

Fear' Bool
Creature causes fear, the Boolean indicates whether the skill is available (True) or used already (False)

LongReach'

Ranged'

Source' Nat Bool
Creature create mana at beginning of turn. Boolean indicates whether the skill is available (True) or used already (False)

Stupid4' Int
The turn, at 0, 1, or 2 not stupide; at 3 stupid; then back to 0

Terror' Bool
Creature causes terror, the Boolean indicates whether the skill is available (True) or used already (False)

Unique'

isStupid :: SkillCore -> Bool

data Pack

Constructors

Pack

Fields

skill :: Skill

text :: String

title :: String

lift :: SkillCore -> Skill

unlift :: Skill -> SkillCore

Because this function uses default values, it is NOT harmless! Use only when initializing data.