Skill

Description

A spinoff of Card |

Documentation

data T blow drawCard fame fear green growth regen slow source stupid terror

The type of skills. Highly polymorphic, because we use two different instances. Callers should only use the concrete instances Skill and State.

Constructors

Ace

Assassin
At turn start, moves against ponent ennemy

Blow blow

Brainless
Moves to random empty spot at turn start

BreathIce

Charge

Discipline

DrawCard drawCard

Fame fame
Contributes to score at turn beginning

Fear fear
Creature causes fear

GreenAffinity green
Creates a forest at turn beginning

Growth growth
+1 hp at turn beginning

Imprecise
Creature attacks a random spot

King

Knight
Identifier of knight

LongReach
Can attack 2 cells away

Powerful
Undealt damage contributes to score

Ranged

Regeneration regen
Regenerates the number of hitpoints at beginning of turn

Sadism
Upon a kill, killed neighbors get -1 attack

Slow slow

1 attack the first turn

StrengthPot
Strength potion in action

Source source
Creature creates mana at beginning of turn

Squire
Knight in front (if any) gains +1 hp and +1 attack upon arrival

Stupid4 stupid

Support

Sylvan
+1 hp/+1 attack when in a forest

Terror terror
Creature causes terror

Unique

Veteran
Immune to fear and terror

Zealot
Immune to fear

type Skill = T () () Nat () () () Nat () Nat () ()

type State = T Bool Bool (Nat, Bool) Bool Bool Bool Nat Bool (Nat, Bool) Nat Bool

isStupid :: State -> Bool

data Pack

Constructors

Pack

Fields

skill :: Skill

text :: String

title :: String

lift :: State -> Skill

unlift :: Skill -> State

Because this function uses default values, it is NOT harmless! Use only when initializing data.