Update

Documentation

data MultiPlayerLobbyAction

Constructors

LobbyUpdateUsername MisoString

LobbySubmitUsername

LobbyInviteUser UserName

CancelInvitationClicked

AcceptInvitationClicked

RejectInvitationClicked

LobbyServerMessage OutMessage

data SinglePlayerLobbyAction

Constructors

LobbySelectTeam (Maybe Team)

data DnDAction a

Constructors

DragEnd
Dragging card in hand ends. When a successful drop is done, this event is fired right after Drop. We rely on that. If Drop was fired last, we would miss it. Be careful on untested browsers.

DragStart HandIndex
Dragging card in hand

Drop
In GameView, this can play the GamePlayEvent Place

DragEnter a

DragLeave a

data GameAction

Actions that are raised by GameView

Constructors

GameDnD (DnDAction Target)
A drag an drop event

GamePlay Event
Play some game event. It can be an event scheduled by the AI Or a EndTurn GamePlayEvent scheduled because the player pressed "End Turn".

GameDrawCards [DrawSource]
Turn was updated previously by GameIncrTurn, time to draw cards from the stack. Then the handler of this event will take care of giving the player control back. This event is translated to a list of events, iteratively consuming the list.

GameIncrTurn
All actions have been resolved, time to update the turn widget and to schedule GameDrawCard. This does NOT translate to a GamePlayEvent.

GameEndTurnPressed
End Turn button pressed in turn widget. For player, schedule attacks then GameIncrTurn; for AI, compute its actions, schedule them, and then schedule attack and GameIncrTurn.

GameInHandMouseEnter HandIndex
Starting hovering card in hand

GameInHandMouseLeave HandIndex
Ending hovering card in hand

GameInPlaceMouseEnter Target
Starting hovering a target

GameInPlaceMouseLeave Target
Ending hovering a target

GameExecuteCmd
Execute a command (dev mode only)

GameUpdateCmd MisoString
Update the command to execute soon (dev mode only)

data LootAction

Actions internal to LootView

Constructors

DeckTo Direction
TODO @smelc delete me

data WelcomeDestination

To which page to go to, from the welcome page

Constructors

MultiPlayerDestination

SinglePlayerDestination

type DeckViewInput = [Card Core]

data Action

Sum type for application events. If drop stuff doesn't work | think whether it's affected by https://github.com/dmjio/miso/issues/478

Constructors

DeckBack
Leave DeckView, go to another view

DeckGo DeckViewInput
Leave a view, go to DeckView

GameAction' GameAction

LootAction' LootAction
Actions internal to LootView

LootGo
Go to LootView

NoOp

MultiPlayerLobbyAction' MultiPlayerLobbyAction

SayHelloWorld

SinglePlayerLobbyAction' SinglePlayerLobbyAction

SinglePlayerBack

SinglePlayerGo

SceneAction' SceneAction

Keyboard (Set Int)

WelcomeGo WelcomeDestination

data SceneAction

Constructors

StepScene

PauseOrResumeSceneForDebugging

JumpToFrameForDebugging Int

logUpdates :: (Monad m, Eq a, ToExpr a) => (Action -> a -> m a) -> Action -> a -> m a

class Interactable m t mseq | m -> t, m -> mseq where

Class defining the behavior of updating m w.r.t. Interaction and DnDAction

Methods

considerAction :: m -> DnDAction t -> Bool

drop :: m -> PlayerSpot -> HandIndex -> t -> mseq

What to do when dropping card at given index, on the given t target The spot indicates the player whose card is being played

getPlayingPlayer :: m -> PlayerSpot

stopWrongDrop :: m -> Bool

When to stop a dropping action

updateDefault :: m -> Interaction t -> mseq

The default behavior

withInteraction :: m -> Interaction t -> m

How to update the abstract state m with the given Interaction

act :: forall m t mseq. Interactable m t mseq => m -> DnDAction t -> Interaction t -> mseq

type GameActionSeq = [(Int, GameAction)]

Event to fire after the given delays (in seconds). Delays add up.

playOne :: GameModel -> Event -> (GameModel, GameActionSeq)

updateGameModel :: GameModel -> GameAction -> Interaction Target -> (GameModel, GameActionSeq)

We MUST return GameAction as the second element (as opposed to GamePlayEvent). This is used for GameIncrTurn for example. The second element must only contain delayed stuff, it's invalid to return a list whose first element has 0. If we did that, we would need to play this event right away, and then we would have new delayed actions to consider which we wouldn't know how to merge with the first ones. If you feel like adding a 0 event, you instead need to play this event right away by doing a recursive call (or use playOne)

updateGameIncrTurn :: Members '[State SharedModel, State (Board Core), State (Board UI)] eff => GameModel -> Eff eff (Either Text (GameModel, GameActionSeq))

This function uses Eff to have multiple State for different types This allows to use put/get to track updates to the state, instead of having variables with increasing version numbers (ticks)

updateLootModel :: LootAction -> LootModel -> LootModel

updateSinglePlayerLobbyModel :: SinglePlayerLobbyAction -> SinglePlayerLobbyModel -> SinglePlayerLobbyModel

updateMultiPlayerLobbyModel :: MultiPlayerLobbyAction -> MultiPlayerLobbyModel -> Effect Action MultiPlayerLobbyModel

Updates a MultiPlayerLobbyModel

delayActions :: m -> [(Int, Action)] -> Effect Action m

toSceneModel :: Scene () -> SceneModel

toSecs :: Int -> Int

Seconds to scheduling delay

tenthToSecs :: Int -> Int

Tenth of seconds to scheduling delay

updateSceneModel :: SceneAction -> SceneModel -> Effect Action SceneModel

updateModel :: Action -> Model -> Effect Action Model

Updates model, optionally introduces side effects | This function delegates to the various specialized functions | and is the only one to handle page changes. A page change event | is one that takes a model from a page and returns a model of another page. | Such an event cannot be dealt with a specialized function | (SpecializedAction -> SpecializedModel -> SpecializedModel), | it needs to be in `Action -> Model -> Model`.

level0GameModel :: Difficulty -> SharedModel -> Teams Team -> GameModel

The initial model, appropriately shuffled with SharedModel rng

levelNGameModel

Arguments

:: Difficulty

-> SharedModel

-> Level

-> Teams (Team, [Card Core])
The initial decks

-> GameModel

A model that takes the decks as parameters, for use after the initial start of the game

unsafeInitialGameModel

Arguments

:: Difficulty

-> SharedModel

-> Teams (Team, [Card Core])
The initial decks

-> Board Core
The board

-> GameModel

An initial model, with a custom board

initialWelcomeModel :: SharedModel -> WelcomeModel

Orphan instances

ToExpr StdGen

Instance details

Methods

toExpr :: StdGen -> Expr

listToExpr :: [StdGen] -> Expr

ToExpr TileUI

Instance details

Methods

toExpr :: TileUI -> Expr

listToExpr :: [TileUI] -> Expr

ToExpr Tile

Instance details

Methods

toExpr :: Tile -> Expr

listToExpr :: [Tile] -> Expr

ToExpr Filepath

Instance details

Methods

toExpr :: Filepath -> Expr

listToExpr :: [Filepath] -> Expr

ToExpr ID

Instance details

Methods

toExpr :: ID -> Expr

listToExpr :: [ID] -> Expr

ToExpr Item

Instance details

Methods

toExpr :: Item -> Expr

listToExpr :: [Item] -> Expr

ToExpr Neutral

Instance details

Methods

toExpr :: Neutral -> Expr

listToExpr :: [Neutral] -> Expr

ToExpr CreatureID

Instance details

Methods

toExpr :: CreatureID -> Expr

listToExpr :: [CreatureID] -> Expr

ToExpr CreatureKind

Instance details

Methods

toExpr :: CreatureKind -> Expr

listToExpr :: [CreatureKind] -> Expr

ToExpr SkillPack

Instance details

Methods

toExpr :: SkillPack -> Expr

listToExpr :: [SkillPack] -> Expr

ToExpr SkillCore

Instance details

Methods

toExpr :: SkillCore -> Expr

listToExpr :: [SkillCore] -> Expr

ToExpr Skill

Instance details

Methods

toExpr :: Skill -> Expr

listToExpr :: [Skill] -> Expr

ToExpr Team

Instance details

Methods

toExpr :: Team -> Expr

listToExpr :: [Team] -> Expr

ToExpr TimedFrame

Instance details

Methods

toExpr :: TimedFrame -> Expr

listToExpr :: [TimedFrame] -> Expr

ToExpr Frame

Instance details

Methods

toExpr :: Frame -> Expr

listToExpr :: [Frame] -> Expr

ToExpr Actor

Instance details

Methods

toExpr :: Actor -> Expr

listToExpr :: [Actor] -> Expr

ToExpr Element

Instance details

Methods

toExpr :: Element -> Expr

listToExpr :: [Element] -> Expr

ToExpr ActorState

Instance details

Methods

toExpr :: ActorState -> Expr

listToExpr :: [ActorState] -> Expr

ToExpr Direction

Instance details

Methods

toExpr :: Direction -> Expr

listToExpr :: [Direction] -> Expr

ToExpr Level

Instance details

Methods

toExpr :: Level -> Expr

listToExpr :: [Level] -> Expr

ToExpr SharedModel

Instance details

Methods

toExpr :: SharedModel -> Expr

listToExpr :: [SharedModel] -> Expr

ToExpr PlayerSpot

Instance details

Methods

toExpr :: PlayerSpot -> Expr

listToExpr :: [PlayerSpot] -> Expr

ToExpr HandIndex

Instance details

Methods

toExpr :: HandIndex -> Expr

listToExpr :: [HandIndex] -> Expr

ToExpr InPlaceEffects

Instance details

Methods

toExpr :: InPlaceEffects -> Expr

listToExpr :: [InPlaceEffects] -> Expr

ToExpr InPlaceEffect

Instance details

Methods

toExpr :: InPlaceEffect -> Expr

listToExpr :: [InPlaceEffect] -> Expr

ToExpr DeathCause

Instance details

Methods

toExpr :: DeathCause -> Expr

listToExpr :: [DeathCause] -> Expr

ToExpr CardSpot

Instance details

Methods

toExpr :: CardSpot -> Expr

listToExpr :: [CardSpot] -> Expr

ToExpr Target

Instance details

Methods

toExpr :: Target -> Expr

listToExpr :: [Target] -> Expr

ToExpr Difficulty

Instance details

Methods

toExpr :: Difficulty -> Expr

listToExpr :: [Difficulty] -> Expr

ToExpr Turn

Instance details

Methods

toExpr :: Turn -> Expr

listToExpr :: [Turn] -> Expr

ToExpr Model

Instance details

Methods

toExpr :: Model -> Expr

listToExpr :: [Model] -> Expr

ToExpr LootModel

Instance details

Methods

toExpr :: LootModel -> Expr

listToExpr :: [LootModel] -> Expr

ToExpr DeckModel

Instance details

Methods

toExpr :: DeckModel -> Expr

listToExpr :: [DeckModel] -> Expr

ToExpr InvitedActorState

Instance details

Methods

toExpr :: InvitedActorState -> Expr

listToExpr :: [InvitedActorState] -> Expr

ToExpr InvitationActorState

Instance details

Methods

toExpr :: InvitationActorState -> Expr

listToExpr :: [InvitationActorState] -> Expr

ToExpr MultiPlayerLobbyModel

Instance details

Methods

toExpr :: MultiPlayerLobbyModel -> Expr

listToExpr :: [MultiPlayerLobbyModel] -> Expr

ToExpr MultiPlayerLobbyError

Instance details

Methods

toExpr :: MultiPlayerLobbyError -> Expr

listToExpr :: [MultiPlayerLobbyError] -> Expr

ToExpr SinglePlayerLobbyModel

Instance details

Methods

toExpr :: SinglePlayerLobbyModel -> Expr

listToExpr :: [SinglePlayerLobbyModel] -> Expr

ToExpr WelcomeModel

Instance details

Methods

toExpr :: WelcomeModel -> Expr

listToExpr :: [WelcomeModel] -> Expr

ToExpr SceneModel

Instance details

Methods

toExpr :: SceneModel -> Expr

listToExpr :: [SceneModel] -> Expr

ToExpr PlayingMode

Instance details

Methods

toExpr :: PlayingMode -> Expr

listToExpr :: [PlayingMode] -> Expr

ToExpr BuildModel

Instance details

Methods

toExpr :: BuildModel -> Expr

listToExpr :: [BuildModel] -> Expr

ToExpr GameModel

Instance details

Methods

toExpr :: GameModel -> Expr

listToExpr :: [GameModel] -> Expr

ToExpr HandFiddle

Instance details

Methods

toExpr :: HandFiddle -> Expr

listToExpr :: [HandFiddle] -> Expr

ToExpr Hovering

Instance details

Methods

toExpr :: Hovering -> Expr

listToExpr :: [Hovering] -> Expr

ToExpr (Card Core)

Instance details

Methods

toExpr :: Card Core -> Expr

listToExpr :: [Card Core] -> Expr

ToExpr (Card UI)

Instance details

Methods

toExpr :: Card UI -> Expr

listToExpr :: [Card UI] -> Expr

ToExpr (CardCommon Core)

Instance details

Methods

toExpr :: CardCommon Core -> Expr

listToExpr :: [CardCommon Core] -> Expr

ToExpr (CardCommon UI)

Instance details

Methods

toExpr :: CardCommon UI -> Expr

listToExpr :: [CardCommon UI] -> Expr

ToExpr (ItemObject Core)

Instance details

Methods

toExpr :: ItemObject Core -> Expr

listToExpr :: [ItemObject Core] -> Expr

ToExpr (ItemObject UI)

Instance details

Methods

toExpr :: ItemObject UI -> Expr

listToExpr :: [ItemObject UI] -> Expr

ToExpr (NeutralObject Core)

Instance details

Methods

toExpr :: NeutralObject Core -> Expr

listToExpr :: [NeutralObject Core] -> Expr

ToExpr (NeutralObject UI)

Instance details

Methods

toExpr :: NeutralObject UI -> Expr

listToExpr :: [NeutralObject UI] -> Expr

ToExpr (Creature Core)

Instance details

Methods

toExpr :: Creature Core -> Expr

listToExpr :: [Creature Core] -> Expr

ToExpr (Creature UI)

Instance details

Methods

toExpr :: Creature UI -> Expr

listToExpr :: [Creature UI] -> Expr

ToExpr (Board Core)

Instance details

Methods

toExpr :: Board Core -> Expr

listToExpr :: [Board Core] -> Expr

ToExpr (Board UI)

Instance details

Methods

toExpr :: Board UI -> Expr

listToExpr :: [Board UI] -> Expr

ToExpr (PlayerPart Core)

Instance details

Methods

toExpr :: PlayerPart Core -> Expr

listToExpr :: [PlayerPart Core] -> Expr

ToExpr (PlayerPart UI)

Instance details

Methods

toExpr :: PlayerPart UI -> Expr

listToExpr :: [PlayerPart UI] -> Expr

ToExpr (Dragging Target)

Instance details

Methods

toExpr :: Dragging Target -> Expr

listToExpr :: [Dragging Target] -> Expr

ToExpr (Interaction Target)

Instance details

Methods

toExpr :: Interaction Target -> Expr

listToExpr :: [Interaction Target] -> Expr