WorldView

Description

Module displaying the world map, which is the first view in legendary edition as well as the view between levels.

Documentation

viewWorldModel :: World -> Styled (View Action)

legendDiv :: View a

The div showing the legend for the character

cellsHeight :: Int

The height of the view, in number of cells

cellsWidth :: Int

The width of the view, in number of cells

mkModel :: Model -> World

both :: Bifunctor p => (a -> d) -> p a a -> p d d

shift :: T -> World -> World

Shifts the world's position by the given direction, if possible

fightSpots :: Map Team [Coord]

The position of fights, hardcoded yes

chooseTeamSpots :: Map Team Coord

The position where to choose the team

mkEncounters :: Bool -> Coord -> Coord -> Map Coord Encounter

encounterToTile :: Encounter -> Tile